fix(cli): split the OTEL event handler out of exec.go, back under the 800-line gate
cmd/ailang/exec.go crossed the check-file-sizes ceiling (807 > 800) in 8c41d41d4,
turning the REQUIRED `test` context red on dev and blocking every open PR.

spanningEventHandler and its streaming callbacks move byte-for-byte to
cmd/ailang/exec_events.go. Pure move: the exec.go diff is 193 deletions and zero
additions, no function body, signature, visibility or import-path change.
exec.go 807 -> 614; exec_events.go 209.
