+++ changelogs/v0.32-current.md
@@ -4,6 +4,19 @@
 
 ## [Unreleased]
 
+### Fixed — `cmd/ailang/exec.go` crossed the 800-line gate and reddened a REQUIRED context
+
+`8c41d41d4` ("chore: checkpoint existing z.ai provider and benchmark work") added a net 9 lines to
+`cmd/ailang/exec.go` (10 insertions, 1 deletion), taking it 798 → **807** and failing
+`make check-file-sizes` (>800). That step
+lives in CI's `test` job, which is a REQUIRED context, so a file-size overrun on an unrelated
+checkpoint blocked every open pull request in the repo.
+
+- `spanningEventHandler` and its streaming callbacks (`newSpanningEventHandler`, `SetContext`,
+  `OnTurnStart`, `OnText`, `OnToolUse`, `OnToolResult`, `OnTurnEnd`, `OnError`) move **byte-for-byte**
+  into `cmd/ailang/exec_events.go`. `exec.go` 807 → 614, new file 209; the diff of `exec.go` is 193
+  deletions and **zero** additions. Same package, no signature, visibility or import-path change.
+
 ### Added — the fleet can now see how much of a subscription bucket it has spent
 
 `M-QUOTA-RATIONING-ROUTING` M2. Routing could ask "is this lane up?" and never "can it afford to
